Switched to java.time libraries and removed useless MyDuration class
Calendar objects are now represented as LocalDateTime objects which allow far more sensible operations on dates.
It should be now way easier to generate a query to add newly created Event objects to the database and back, the query itself should be done (untested) but code for PreparedStatement is still missing.